Isolate settings in a separate shared module to avoid pulling in react as a dependency in the server
