+++ common/settingsEnum.js
@@ -1,0 +1,7 @@
+export default {
+  BOOLEAN: 'boolean',
+  LIST: 'list',
+  NODE: 'node',
+  NUMBER: 'number',
+  STRING: 'string'
+};
